Move FluxMaps to_hdulist and from_hdulist
